+++ CHANGELOG.md
@@ -27,6 +27,7 @@ The releases of Vieb aim to follow [semantic versioning](https://semver.org).
 - Command mapping list to the header of commands on the help page
 - Setting "menupage" to control when the context menu of the webpage should appear, possibly over the website's menu
 - Setting "menuvieb" to control when the context menu should appear for the Vieb elements like the navbar and the tabbar
+- Pointer mode actions for most context menu options, such as "pointer.downloadAudio", "pointer.newtabVideo" and "pointer.copyImageBuffer"
 
 ### Changed